IndexedDB/idbfactory_open13.htm

<!DOCTYPE html>
<meta charset="utf-8">
<title>IDBFactory.open() - no argument</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://dvcs.w3.org/hg/IndexedDB/raw-file/tip/Overview.html#widl-IDBFactory-open-IDBOpenDBRequest-DOMString-name-unsigned-long-long-version">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>
<div id="log"></div>
<script>
    test( function() {
        assert_throws({name: 'TypeError'}, function() {
            indexedDB.open();
        });
    }, "IDBFactory.open() - no argument.");
</script>
